requirements.txt
# Jasper AI v3 — Requirements (all free, all local)

# Core
streamlit>=1.32.0
chromadb>=0.4.24
sentence-transformers>=2.7.0
requests>=2.31.0

# File ingestion
pdfplumber>=0.11.0
python-docx>=1.1.0

# Knowledge crawler
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.0
arxiv>=2.1.0

# GitHub agent
PyGithub>=2.1.0

# Project video generation
moviepy>=1.0.3
Pillow>=10.0.0

# Encoding fix
colorama>=0.4.6
